#56: add asyncio tests, refactor sync tests
