+++ examples/rtu_client_example.py
@@ -17,71 +17,11 @@ bus address and UART communication speed can be defined by the user.
 
 # import modbus client classes
 from umodbus.serial import ModbusRTU
+from examples.common.register_definitions import register_definitions, setup_callbacks
+from examples.common.rtu_client_common import IS_DOCKER_MICROPYTHON
+from examples.common.rtu_client_common import slave_addr, rtu_pins
+from examples.common.rtu_client_common import baudrate, uart_id, exit
 
-IS_DOCKER_MICROPYTHON = False
-try:
-    import machine
-    machine.reset_cause()
-except ImportError:
-    raise Exception('Unable to import machine, are all fakes available?')
-except AttributeError:
-    # machine fake class has no "reset_cause" function
-    IS_DOCKER_MICROPYTHON = True
-    import json
-
-
-# ===============================================
-# RTU Slave setup
-# act as client, provide Modbus data via RTU to a host device
-# ModbusRTU can get serial requests from a host device to provide/set data
-# check MicroPython UART documentation
-# https://docs.micropython.org/en/latest/library/machine.UART.html
-# for Device/Port specific setup
-#
-# RP2 needs "rtu_pins = (Pin(4), Pin(5))" whereas ESP32 can use any pin
-# the following example is for an ESP32.
-# For further details check the latest MicroPython Modbus RTU documentation
-# example https://micropython-modbus.readthedocs.io/en/latest/EXAMPLES.html#rtu
-rtu_pins = (25, 26)         # (TX, RX)
-slave_addr = 10             # address on bus as client
-baudrate = 9600
-uart_id = 1
-
-try:
-    from machine import Pin
-    import os
-    from umodbus import version
-
-    os_info = os.uname()
-    print('MicroPython infos: {}'.format(os_info))
-    print('Used micropthon-modbus version: {}'.format(version.__version__))
-
-    if 'pyboard' in os_info:
-        # NOT YET TESTED !
-        # https://docs.micropython.org/en/latest/library/pyb.UART.html#pyb-uart
-        # (TX, RX) = (X9, X10) = (PB6, PB7)
-        uart_id = 1
-        # (TX, RX)
-        rtu_pins = (Pin(PB6), Pin(PB7))     # noqa: F821
-    elif 'esp8266' in os_info:
-        # https://docs.micropython.org/en/latest/esp8266/quickref.html#uart-serial-bus
-        raise Exception(
-            'UART0 of ESP8266 is used by REPL, UART1 can only be used for TX'
-        )
-    elif 'esp32' in os_info:
-        # https://docs.micropython.org/en/latest/esp32/quickref.html#uart-serial-bus
-        uart_id = 1
-        rtu_pins = (25, 26)             # (TX, RX)
-    elif 'rp2' in os_info:
-        # https://docs.micropython.org/en/latest/rp2/quickref.html#uart-serial-bus
-        uart_id = 0
-        rtu_pins = (Pin(0), Pin(1))     # (TX, RX)
-except AttributeError:
-    pass
-except Exception as e:
-    raise e
-
-print('Using pins {} with UART ID {}'.format(rtu_pins, uart_id))
 
 client = ModbusRTU(
     addr=slave_addr,        # address on bus
@@ -95,67 +35,15 @@ client = ModbusRTU(
 )
 
 if IS_DOCKER_MICROPYTHON:
+    import json
     # works only with fake machine UART
     assert client._itf._uart._is_server is True
-
-
-def reset_data_registers_cb(reg_type, address, val):
-    # usage of global isn't great, but okay for an example
-    global client
-    global register_definitions
-
-    print('Resetting register data to default values ...')
-    client.setup_registers(registers=register_definitions)
-    print('Default values restored')
-
-
-# common slave register setup, to be used with the Master example above
-register_definitions = {
-    "COILS": {
-        "RESET_REGISTER_DATA_COIL": {
-            "register": 42,
-            "len": 1,
-            "val": 0
-        },
-        "EXAMPLE_COIL": {
-            "register": 123,
-            "len": 1,
-            "val": 1
-        }
-    },
-    "HREGS": {
-        "EXAMPLE_HREG": {
-            "register": 93,
-            "len": 1,
-            "val": 19
-        }
-    },
-    "ISTS": {
-        "EXAMPLE_ISTS": {
-            "register": 67,
-            "len": 1,
-            "val": 0
-        }
-    },
-    "IREGS": {
-        "EXAMPLE_IREG": {
-            "register": 10,
-            "len": 1,
-            "val": 60001
-        }
-    }
-}
-
-# alternatively the register definitions can also be loaded from a JSON file
-# this is always done if Docker is used for testing purpose in order to keep
-# the client registers in sync with the test registers
-if IS_DOCKER_MICROPYTHON:
     with open('registers/example.json', 'r') as file:
-        register_definitions = json.load(file)
+        register_definitions = json.load(file)  # noqa: F811
+
 
 # reset all registers back to their default value with a callback
-register_definitions['COILS']['RESET_REGISTER_DATA_COIL']['on_set_cb'] = \
-    reset_data_registers_cb
+setup_callbacks(client, register_definitions)
 
 print('Setting up registers ...')
 # use the defined values of each register type provided by register_definitions
@@ -177,3 +65,5 @@ while True:
         print('Exception during execution: {}'.format(e))
 
 print("Finished providing/accepting data as client")
+
+exit()
